refine handling of local files
rather than just sticking it all in the caches folder, local files are now stored as documents (i.e. not auto-purgeable) and moved to the caches folder when they would otherwise merit deletion, in order to quickly be restored if e.g. switching accounts or changing which construction sites you can access